+++ extensions/rsync-commands/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Rsync Commands Changelog
 
+## [1.1.0] - 2020-05-21
+- Configure your desired identity file (private key) for SSH sync.
+- Configure custom commands to run before and after the rsync command itself.
+
 ## [Initial Version] - 2022-05-19
 
 - Create, update, delete and duplicate pre-configured rsync commands.